# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Flamoral Dating Platform API
  description: |
    Complete API documentation for the Flamoral Dating Platform.

    This dating platform features:
    - User authentication and profiles
    - Discovery and matching system
    - Real-time messaging
    - Subscription and coins system
    - Boost products
    - Privacy controls
    - Notifications
    - Payment processing
  version: 1.0.0
  contact:
    name: Flamoral Platform Team
    email: platform@flamoral.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.flamoral.com
    description: Production API
...
paths:
...
  /api/v1/subscriptions/cancel:
    post:
      tags: [Subscriptions]
      summary: Cancel subscription
      operationId: cancelSubscription
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subscription cancelled
  /api/v1/subscriptions/reactivate:
    post:
      tags: [Subscriptions]
      summary: Reactivate subscription
      operationId: reactivateSubscription
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subscription reactivated

  # ============== USER SERVICE - COINS ==============
...
  /api/v1/payments/methods:
    get:
      tags: [Payments]
      summary: Get payment methods
      operationId: getPaymentMethods
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Payment methods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentMethod'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    PaymentMethod:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        last4:
          type: string
        brand:
          type: string
        isDefault:
          type: boolean
...